Trasferred couples view in season app and fix URL issues
