+++ tennisblock/season/templates/season/seasons.html
@@ -28,7 +28,7 @@
                         <td>{{ season.enddate }}</td>
                         <td>{{ season.blockstart }}</td>
                         <td>{{ season.blocktime }}</td>
-                        <td><a href="{% url 'couple_editor' season.pk %}">Couple Editor</a></td>
+                        <td><a href="{% url 'season:couple_editor' season.pk %}">Couple Editor</a></td>
                     </tr>
                 {% endfor %}
             </tbody>
