+++ tennisblock/season/views.py
@@ -1,12 +1,14 @@
 from dateutil.parser import parse
 from django.views.generic import TemplateView, CreateView
 from django.urls import reverse_lazy
+from django.shortcuts import get_object_or_404
+from django.db import IntegrityError
 
 from TBLib.view import class_login_required
 
-from blockdb.models import Season, Meeting, Player, SeasonPlayer
+from blockdb.models import Season, Meeting, Player, SeasonPlayer, Couple
 
-from .forms import SeasonForm
+from .forms import SeasonForm, CoupleForm
 from api.apiutils import build_meetings_for_season
 
 @class_login_required
@@ -147,3 +149,67 @@ class SeasonCreate(CreateView):
     success_url = reverse_lazy('season:seasons')
     form_class = SeasonForm
     model = Season
+
+
+@class_login_required
+class CouplesView(TemplateView):
+    template_name = "season/couple_editor.html"
+
+    def get_context_data(self, pk=None, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        season = get_object_or_404(Season, pk=pk)
+        context['season'] = season
+
+        players = SeasonPlayer.objects.filter(season=season)
+        couples = Couple.objects.filter(season=season)
+
+        context['players'] = players
+        context['couples'] = couples
+        initial = {
+            'season': season,
+            'fulltime': False,
+            'blockcouple': True,
+            'canschedule': True
+        }
+        context['form'] = CoupleForm(season, initial=initial)
+
+        return context
+
+    def get(self, request, pk=None, **kwargs):
+        context = self.get_context_data(pk=pk, **kwargs)
+        return self.render_to_response(context)
+
+    def post(self, request, pk=None, **kwargs):
+
+        try:
+            s = get_object_or_404(Season, pk=pk)
+
+            form = CoupleForm(s, request.POST)
+            if form.is_valid():
+                print("Valid Couple form..updating...")
+                Couple.objects.create(
+                    season=s,
+                    name=form.data['name'],
+                    male=form.spguy.player,
+                    female=form.spgal.player,
+                    fulltime=form.data.get('fulltime', 'off') == 'on',
+                    blockcouple=form.data.get('blockcouple', 'off') == 'on',
+                    canschedule=True
+                ).save()
+                print("Inserted couple")
+            else:
+                print("Invalid Couple")
+
+        except IntegrityError:
+            print("Attempt to insert duplicate couple!")
+
+        context = self.get_context_data(pk=pk, **kwargs)
+        return self.render_to_response(context)
+
+    def insert_couple(self, couple=None):
+        """ Check for duplicates before inserting """
+        if not Couple.objects.filter(
+                male=couple.male,
+                female=couple.female, season=couple.season).exists():
+            couple.save()
