+++ tennisblock/webapp/views.py
@@ -1,9 +1,5 @@
-from textwrap import dedent
-
 from django.views.generic import TemplateView
 from django.core.mail import send_mail
-from django.db import IntegrityError
-from django.shortcuts import get_object_or_404
 from django.contrib.auth.mixins import LoginRequiredMixin
 
 from django.forms.formsets import (
@@ -14,10 +10,9 @@ from django.shortcuts import render
 from django.conf import settings
 
 from TBLib.view import TennisLoginView
-from blockdb.models import Season, Couple, SeasonPlayer
 from .forms import ContactForm
 
-from .forms import CoupleForm, AvailabilityForm
+from .forms import AvailabilityForm
 from TBLib.view import class_login_required
 
 
@@ -115,67 +110,3 @@ class ContactView(TemplateView):
         return render(request,
                       self.template_name,
                       {'form': form})
-
-
-@class_login_required
-class CouplesView(TemplateView):
-    template_name = "couple_editor.html"
-
-    def get_context_data(self, pk=None, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        season = get_object_or_404(Season, pk=pk)
-        context['season'] = season
-
-        players = SeasonPlayer.objects.filter(season=season)
-        couples = Couple.objects.filter(season=season)
-
-        context['players'] = players
-        context['couples'] = couples
-        initial = {
-            'season': season,
-            'fulltime': False,
-            'blockcouple': True,
-            'canschedule': True
-        }
-        context['form'] = CoupleForm(season, initial=initial)
-
-        return context
-
-    def get(self, request, pk=None, **kwargs):
-        context = self.get_context_data(pk=pk, **kwargs)
-        return self.render_to_response(context)
-
-    def post(self, request, pk=None, **kwargs):
-
-        try:
-            s = get_object_or_404(Season, pk=pk)
-
-            form = CoupleForm(s, request.POST)
-            if form.is_valid():
-                print("Valid Couple form..updating...")
-                Couple.objects.create(
-                    season=s,
-                    name=form.data['name'],
-                    male=form.spguy.player,
-                    female=form.spgal.player,
-                    fulltime=form.data.get('fulltime', 'off') == 'on',
-                    blockcouple=form.data.get('blockcouple', 'off') == 'on',
-                    canschedule=True
-                ).save()
-                print("Inserted couple")
-            else:
-                print("Invalid Couple")
-
-        except IntegrityError:
-            print("Attempt to insert duplicate couple!")
-
-        context = self.get_context_data(**kwargs)
-        return self.render_to_response(context)
-
-    def insert_couple(self, couple=None):
-        """ Check for duplicates before inserting """
-        if not Couple.objects.filter(
-                male=couple.male,
-                female=couple.female, season=couple.season).exists():
-            couple.save()
